refactor: promotion 결과 조립 AdminPromotionResultAssembler로 통합
